refactor: create home page instead of leaving the gallery in App
